+++ app.rb
@@ -3,14 +3,14 @@
 require 'sinatra/base'
 require 'sinatra/reloader'
 require 'pg'
-require './lib/peeps'
-require './lib/database_connection_setup'
+require 'peeps'
+require 'database_connection_setup'
 # as the connection setup is required here,
 # the connection will be set accordingly when app.rb is started
 
 class Chitter < Sinatra::Base
   configure :development do
-  register Sinatra::Reloader
+    register Sinatra::Reloader
   end
 
   enable :sessions
@@ -21,8 +21,6 @@ class Chitter < Sinatra::Base
   end
 
   post '/mypeeps' do
-    puts params[:new_peep]
-    connection = PG.connect(dbname: 'chitter_test')
     DatabaseConnection.query(
       "INSERT INTO peeps (content, peeper, post_time) VALUES ($1, $2, current_timestamp)",
       [params[:new_peep], 'DEV_TESTING']
@@ -31,10 +29,8 @@ class Chitter < Sinatra::Base
   end
 
   get '/mypeeps' do
-    @result = DatabaseConnection.query(
-      "SELECT * FROM peeps WHERE peeper = $1 ORDER BY post_time DESC",
-      ['DEV_TESTING']
-    )
+    DatabaseConnection.setup('chitter_test')
+    @result = Peeps.show_mine
     erb :'/peeps/mypeeps'
   end
